requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
sqlalchemy
psycopg2-binary
pydantic-settings
prometheus-client
